pyproject.toml
[project]
name = "BelegPilot"
version = "0.1.0"
description = "Intelligent receipt & invoice extraction pipeline"
requires-python = ">=3.12"
license = "MIT"

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.28.0",
    "tenacity>=9.0.0",
    "opencv-python-headless>=4.10.0",
    "pillow>=11.0.0",
    "pytesseract>=0.3.13",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "python-multipart>=0.0.18",
    "structlog>=24.4.0",
    "opentelemetry-api>=1.29.0",
    "opentelemetry-sdk>=1.29.0",
    "opentelemetry-instrumentation-fastapi>=0.50b0",
    "opentelemetry-exporter-otlp>=1.29.0",
    "arize-phoenix[otel]>=5.0.0",
    "prometheus-client>=0.21.0",
    "bcrypt>=4.2.0",
    "python-jose>=3.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "aiosqlite>=0.20.0",
    "httpx[http2]>=0.28.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "SIM", "TCH"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
